refactor network title logic into its own function
